Remove the connect function and let the client connect automatically
I have tested this with a broker; various scenarios where I started
Hare before the broker was online, tried to shutdown the broker while
Hare was up, tried to reconnect, etc. Seems to work, and I think it is
simpler.

Using handle_continue to break up the steps, and listen to the events
emitted from Tortoise when the connection change state.