Restructure the hare module
Create a work queue that stores orders and turns them into effects; if
the work produce a ref we will track that; we will retry orders that
returned an error. If orders are placed while we are offline we will
execute them when we are online again. Some retry mechanisms are in
place, but we probably need to track how many times a given order has
been tried and fail if maximum is reached.