Implement a Hare.reconnect/0 function
This will disconnect from the broker, add the subscriptions for
re-subscription (just to make sure our world view is consistent once
we get a new connection), and add the currently pending messages to
the work list.

This might cause us to have a message delivered twice, but we should
not drop messages.