Add options to the Hare commands that are separate package opts
Some of the MQTT packages takes options, such as what QoS a
subscription should be requested, or what QoS a publish should be
send. This change introduce a work list opts list that can be used to
store options for the work list (separate from the package opts); this
will be useful for implementing time to live (ttl) on commands, and
other options that might be interesting in the future.

Notice that this change just implement the capability of adding the
option list via the public interface and passing it around internally,
currently no part of the system is using the options.