Change the application supervision strategy to rest for one
I will remove the app_handler, and have tortoise send messages to the
hare process; this will require the tortoise handler to know about the
pid of Hare, as we will no longer call the gen_server cast handlers.

The reasoning behind the supervision strategy is described in a
comment in the code, currently as follows:

Supervision strategy is rest for one, as a crash in Hare would result
in inconsistent state in Hare; we would not be able to know about the
subscription state; so we teardown the tortoise if Hare crash. Should
the Hare.Supervisor crash, Hare should re-subscribe to the topic
filters it currently know about, so that should be okay.