Send Tortoise Results directly from Tortoise client to hare process
Changed the handle result callback to a message pass; the tortoise
client needed to know about the hare pid, so this is now part of its
state as well. Notice, because of the way we supervise the processes,
this should be fine; if the Hare process crashes we would restart the
Tortoise Client as well.