Replace the app handler with a "Hare handler"
This commit introduces a hare handler, simply called handler, which
implements a behaviour with four callbacks:

- connection/1 :: called when the tortoise connection changes status
- subscription/2 :: called when a subscription changes status
- handle_message/2 :: called when we receive a message
- handle_error/1 :: called when errors occur in the system

Errors can be message delivery failures and such. All, but the
handle_message, callbacks are optional. The project implement a
Hare.Handler.Logger, which will log all the messages it sees, has been
provided; in real projects this would get replaced with an application
specific callback handler.