Fix topic return type for :payload_decode_error
The `topic` returned in `payload_decode_error` is the same "topic" in
`Tortoise.Handler.handle_message/3` which means it is a list of topic
levels, but the type for `Jackalope.Handler.handle_message/2` topic is a
string.

This fixes that typing to match.

see https://github.com/smartrent/jackalope/blob/main/lib/jackalope/tortoise_handler.ex#L58