Add Jackalope.Session connection_status/0
Jackalope provides connection status to the implementation of Jackalope.Handler when it changes. However, some situations may require the implementation (or another process) to query the current connection status (i.e. the implementation may be a process that crashed and lost state and needs to rebuild it without waiting for Jackalope connection status to change for the needed event)